Reserve logic changed
The reserve logic is now changed. In fact, now the first thing the
tool does is sending the reserve request. Then, we use several checks
to control the validity of the reserve we made (e.g. correct internal
version, no minor reserve on major drafts...)
Based on the checks we perform some actions, and then
the process of retrieving the Log document related
to the reserve request starts. The process can be stopped whenever
we want, in fact it will restart as we login into the dcf again.
The important thing is that now the process is perfectly integrated
without code duplication.
The logic and the code are much easier now.

Some bugs were fixed.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>