Progress bar added
A progress bar showing the import process of new catalogue
version during the reserve process was added.
This progress bar is not modal, so the application is not
blocked by it, it is useful to show to the user that the
application is doing something.
During the import process the closure of the application
is disabled to prevent data corruption.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>