Reserve process fixed
The reserve process was ordered in a better way. Order:
 1- reserve request is sent
 2- log is retrieved (blocking step)
 3- checks are made on the log, we check if the reserve was successful,
     if we had sent a minor reserve for a draft major catalogue (this
     is an error) and if we have the last internal version of the catalogue
     or not.
 4- based on the checks we warn the user, and the last internal version
   of the catalogue is downloaded if needed

If in 2 we find the dcf busy for a reserve operation,
the editing is forced and we can edit at
our own risk. When the dcf gives a response, we notify the user if the
reserve was Ok or not. If not ok, it blocks the editing and warn the
user that he must merge the changes with the new catalogue.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>