Fixed forced editing
If a catalogue is forced to being edited a copy of the database is
created with a new version (x.y.z.FORCED). If then the reserve
goes well, a new catalogue with the same data of the forced one
is created. If the reserve failed, then we block the editing of the
FORCED version and block also the upload.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>