Import process improved
The import process was subdivided in several classes in order to improve
its modularity. In particular:
- CatalogueSheetImporter imports the catalogue sheet
- AttributeSheetImporter imports the attributes sheet
- HierarchySheetImporter imports the hierarchies sheet
- TermSheetImporter imports the terms sheet (only term information)

- TermAttributeImporter imports the values of attributes contained in
the term sheet (identified by the attributes names)

- ParentImporter imports the parent-child relationships contained in
the term sheet (identified by the hierarchies code + flag, parentCode,
reportable, hierarchyCode, order)

- NotesSheetImporter imports the release notes of the catalogues

The base class of all these classes is SheetImporter.java which implements
the common steps used in all the above classes.

Moreover, also the TermTypeImporter was created to import the term types
of the catalogue (if present). Note that this class does not extend
the SheetImporter base class, because it has a different logic.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>