Drag and drop bug fix
The drag and drop procedure was fixed. In fact, the previous version
usually generates negative order integer for terms.
Moreover, the order integer of terms is now normalized, that is,
the integer series of natural numbers is respected. E.g.
if the orders are 1,4,5,7,8,10 the terms are indeed ordered but
the numbers are not so good. So the procedure convert them into
1,2,3,4,5,6. This is applied both for drag n drop and for cut/copy/paste
operations.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>