Bug fix and Error management
The delete catalogue bug was fixed, now it always works.
Moreover, an additional warning message is added if a catalogue is
downloaded from the dcf but no attachment is found. In this case,
the download process is stopped and the user is warned. If the
catalogue CATUSERS cannot be downloaded the default user access level
is set to DATA PROVIDER.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>