Append function
1) The import excel process was enhanced in order to accomodate also
the possibility of appending new terms.
In fact, it is now possible to define temporary terms by giving them
a code like: TEMP_ + whatever we want. Note that an assumption of the
program is that the temporary codes are unique among them, therefore
be careful! Examples of temporary codes are TEMP_0, TEMP_1234, TEMP_salad,
TEMP_123_ribes.

Note that it is also possible to define the parent code
of a temporary term with another temporary term. In fact, if you want
to give to TEMP_1 the parent TEMP_parent_1 you can do it by
setting as TEMP_1.parentCode the value TEMP_parent_1 in the
hierarchy you want.

IMPORTANT NOTE: if flag, order or reportable are left blank, they
are automatically inferred. In fact, the flag is by default set to 1.
The term is assumed reportable if not specified, and if an order is
not specified, the term will get the first available order code among
siblings under its parent.

Note also that all these assumptions are also valid for non
temporary terms!

2) Solved an import bug which assumed the wrong db path for catalogues.

BUGS:
A bug seems to be present: if we start an unreserve action, sometimes
the Dcf.java is not capable of finding the DefaultProperties.xml file
raising an exception. Note that if the file is not found, the program
will connect by default with the dcf TEST.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>