Dcf Log management
The dcf log is now managed in a buffered way to make the process
more efficient for large logs. Moreover, now the entire log is
saved into a java class (DcfLog.java) which can be used to retrieve
the log information.

If a pending action is performed and some errors occur, then a table
showing all the errors is displayed.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>