Log errors viewer
The entire dcf log related to UploadCatalogueFile actions is now
stored into the DcfLog object. This object is then used to check
the correctness of web service operations and other things.

If an upload catalogue file operation is performed (as a reserve,
unreserve, publish or upload data actions) but some errors occurred,
a form displaying all the dcf log errors is now showed, in order to
help users understanding why the operation is failed.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>