Import parallel
The import process was improved by introducing two threads for
reading and consuming the data. One thread reads the data and
the other writes the read data into the database. This is
done in a cycle, in order to be able to write while
the other thread is reading the next batch of data.

Benchmark on MTX 8.7 (direct import by xlsx, not download):

time for importing the term sheet: around 70 seconds
Performed on EFSA laptop: 8 gb ram,
Processor Intel(R) Core(TM) i3-4000M CPU @ 2.40Ghz
64 bit


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>