Import process improved
Now the parent terms and the term attributes are imported in a parallel
way in order to speed up the process. The batch size of the term
import was also changed and set to 100, since the performance
both in terms of memory used and computational time are better
(previous batch size was 2000).


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>