Link with sas procedure
A new button named "Create .xml changes file" was created under the
tools menu. This button allows sending the exported version of
the current catalogue (.xlsx format) to the sas server.
In particular, a .start file is created first and then, as soon as the
copy is completed, it is renamed with the .xlsx extension (in order to
avoid file system problems due to latencies). Moreover, a .end file
is created in the end to warn the sas server that the .xlsx file is
ready to be used.
The sas server will recognize the .end file and then it will convert
it into an .xml file which contains the differences between the loaded
.xlsx catalogue and its official version, in terms of dcf operations
(in order to be able to upload them with a single UploadCatalogueFile
operation).

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>