Default hierarchy bug fix
The default hierarchy is now correctly displayed and refreshed.
Moreover, the HierarchySelector was adapted to the default hierarchy.

(The default hierarchy is the hierarchy which is shown to read only
users when we do not know which hierarchy to show in the main panel,
in particular,
the default "default hierarchy" is the master hierarchy, but it can be
changed) Note that if a default hierarchy is set and the user is
a read only one, the master hierarchy will be hidden.

Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>