Sas procedure improved
The communication with the sas procedure was enhanced. In particular,
now the .xml which is created into the remote folder can be downloaded
using the XmlChangesCreator.java (it uses the polling technique to
wait the file). If a new xml creation is requested while an old one was
still on going but not processed, the old request is now deleted by
removing the .xlsx file and .end file on the remote sas folder.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>