Enhanced Log & Fixed bug
The log of upload catalogue file actions is now enhanced, in fact,
also the validation errors node are now showed to the user.
The DcfResponse was extended adding the ERR status, since it is another
possible status returned from the dcf.
General bug fixed (deletion of temporary files and stuff like that)