Download xml process
If upload data is started, first a download process starts to move into
the local folder the .xml updates file generated by the sas procedure.
Then the upload of that .xml is done through an upload catalogue file
action.
Now if the download process is started (i.e. we loop until we find the
.xml file) and the application is closed before the upload cat file is
performed, the tools saves the download process as an additional pending
action, in order to restart the download when the application restarts.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>