Double screen bug fix
If:
- two screen are used
- a window is moved to non main screen
- the window is closed in the non main screen (causing the save of the
window position and dimensions)
- the tool is subsequently run on a single screen mode and the same
window as before is opened

Then:
The window will appear outside of the screen since its position was
saved with the double screen mode

This bug was fixed adjusting the window position to fit the single
screen dimensions if a single screen is used. If multiple screens
are used then no adjustment is done.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>