Not in use hierarchies and default hierarchy features
It is now possible to set a subset of hierarchies
which are no more in use. These hierarchies will not be shown
to the users (except for catalogue managers).

It is now possible to set a default hierarchy, which is the
hierarchy that is first shown to users when the catalogue is
opened for the first time, and in general it is the main
hierarchy.

Syntax:

To use these features you should insert in the scope notes
of the catalogue the following things:

- Default hierarchy:
“catalogue scope notes$[defaultHierarchy=hierarchyCode]”

- Not in use hierarchies:
“catalogue scope notes$[notUsedHierarchies=hierarchyCode1,hierarchyCode2,…]”

Note that you can also combine the two features, e.g.:

"catalogue scope notes$[defaultHierarchy=hierarchyCode][notUsedHierarchies=hierarchyCode1,hierarchyCode2,…]”



Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>