Export order fixed, login fixed
The order of elements in the export excel is now fixed according to
the order parameter for hierarchies and attributes and to
term code for terms. Also the hierarchies and the attributes
columns in the terms sheet are ordered according to their order.

The login was fixed. Data providers got errors instead of correct
notifications.


Signed-off-by: AVON Valentino <Valentino.AVON@ext.efsa.europa.eu>